Move implementation out of Sidebar component
